janken-system created

+++ module/talk/voicevox.py
@@ -3,6 +3,9 @@ import io
 import wave
 import pyaudio
 import time
+from moviepy.editor import AudioFileClip
+import tempfile
+import os
 
 
 class Voicevox:
@@ -34,28 +37,44 @@ class Voicevox:
 		# メモリ上で展開
 		audio = io.BytesIO(res.content)
 
-		with wave.open(audio,'rb') as f:
-			# 以下再生用処理
-			p = pyaudio.PyAudio()
+		if res.status_code == 200:
+			# 音声時間取得
+			temp_file = tempfile.NamedTemporaryFile(delete=False, suffix=".mp3")
+			temp_file.write(res.content)
+			temp_file.close()
 
-			def _callback(in_data, frame_count, time_info, status):
-				data = f.readframes(frame_count)
-				return (data, pyaudio.paContinue)
+			try:
+				audioFile = AudioFileClip(temp_file.name)
+				# 音声ファイルの長さ（秒）を取得する
+				duration = audioFile.duration
+				print(f"音声ファイルの長さは {duration} 秒です。")
+			finally:
+				os.unlink(temp_file.name)
 
-			stream = p.open(format=p.get_format_from_width(width=f.getsampwidth()),
-							channels=f.getnchannels(),
-							rate=f.getframerate(),
-							output=True,
-							stream_callback=_callback)
+			with wave.open(audio, 'rb') as f:
+				# 以下再生用処理
+				p = pyaudio.PyAudio()
 
-			# Voice再生
-			stream.start_stream()
-			while stream.is_active():
-				time.sleep(0.1)
+				def _callback(in_data, frame_count, time_info, status):
+					data = f.readframes(frame_count)
+					return (data, pyaudio.paContinue)
 
-			stream.stop_stream()
-			stream.close()
-			p.terminate()
+				stream = p.open(format=p.get_format_from_width(width=f.getsampwidth()),
+								channels=f.getnchannels(),
+								rate=f.getframerate(),
+								output=True,
+								stream_callback=_callback)
+
+				# Voice再生
+				stream.start_stream()
+				while stream.is_active():
+					time.sleep(0.1)
+
+				stream.stop_stream()
+				stream.close()
+				p.terminate()
+		else:
+			print("音声の取得に失敗しました。")
 
 
 # def main():
